Tidying up and pushing code places
